+++ tests/test_invocation.py
@@ -5,7 +5,7 @@ from uuid import uuid4
 from galileo_core.schemas.protect.response import Response
 from pytest import mark
 
-from galileo_protect.invoker import ainvoke, invoke
+from galileo_protect.invocation import ainvoke, invoke
 from galileo_protect.langchain import ProtectTool
 from galileo_protect.schemas import Payload, Ruleset
 from tests.data import A_PROTECT_INPUT, A_STAGE_NAME
